Fix reconnect black screen: Two-sided session-ID handshake (Protocol v2)
Root cause: Windows-side stale-ACK filtering created DEADLOCK on reconnect.
Windows session 2 rejected EVERY ACK as 'Duplicate ACK for session 1 ignored'
because Android had no concept of session ID - it kept sending ACKs that Windows
attributed to the old session. android_ready_ never became true for session 2,
so StreamLoop never started, Android got 0 frames (black screen).

Fix (Protocol v2 - both Windows AND Android changes required):

Windows side:
- Protocol.h: Bump to PROTOCOL_VERSION = 2, document 16-byte stream_info format
- Session.h: session_id_ changed to uint16_t (was uint64_t)
- Session.cpp: Send session_id in stream_info bytes 12-13; ACK callback validates
  ack_session_id matches current session_id_
- TouchReceiver.h/cpp: Parse session_id from ACK packet bytes [6-7], pass to
  callback as uint16_t (was std::string sender_ip)

Android side:
- TcpStreamReceiver.kt: Parse session_id from stream_info (type 2 message)
  bytes 12-13, pass to onSenderIp(ip, sessionId) callback
- ConnectionManager.kt: Store currentSessionId, reset on disconnect, pass to
  TouchSender.sendAck(sessionId); update both USB and WiFi onSenderIp callbacks
- TouchSender.kt: sendAck(sessionId) includes session_id in ACK packet bytes [6-7]

Verification:
- Windows logs show session ID incrementing on reconnect (1, 2, 3...)
- Each new session's ACK is accepted and streaming starts immediately
- No 'Stale ACK' spam on successful reconnects

Changed files:
- windows/src/Protocol.h
- windows/src/Session.h, Session.cpp
- windows/src/TouchReceiver.h, TouchReceiver.cpp
- android/app/src/main/java/com/pocketdisplay/app/TcpStreamReceiver.kt
- android/app/src/main/java/com/pocketdisplay/app/TouchSender.kt
- android/app/src/main/java/com/pocketdisplay/app/ConnectionManager.kt
- CLAUDE.md

+++ windows/src/Session.h
@@ -97,7 +97,8 @@ private:
 
     // Per-session ID to detect stale ACKs from previous sessions during reconnect.
     // Each new Session gets a unique ID; ACK callback validates it matches.
-    uint64_t session_id_ = 0;
+    // Protocol v2: session_id is uint16_t on wire (0-65535 range).
+    uint16_t session_id_ = 0;
 
     std::thread stream_thread_;
     std::thread resend_thread_;
